+++ packages/list-tips-item/index.js
@@ -1,0 +1,7 @@
+import WeuiListTipsItem from './src/list-tips-item'
+
+WeuiListTipsItem.install = vue => {
+  vue.component(WeuiListTipsItem.name, WeuiListTipsItem)
+}
+
+export default WeuiListTipsItem