+++ packages/list-tips-item/src/list-tips-item.vue
@@ -1,0 +1,10 @@
+<template>
+  <div class="weui-list-tips__item">
+    <slot></slot>
+  </div>
+</template>
+<script>
+export default {
+  name: 'WeuiListTipsItem'
+}
+</script>