+++ packages/list-tips/index.js
@@ -1,0 +1,7 @@
+import WeuiListTips from './src/list-tips'
+
+WeuiListTips.install = vue => {
+  vue.component(WeuiListTips.name, WeuiListTips)
+}
+
+export default WeuiListTips