+++ packages/list-tips/src/list-tips.vue
@@ -1,0 +1,10 @@
+<template>
+  <div class="weui-list-tips">
+    <slot></slot>
+  </div>
+</template>
+<script>
+export default {
+  name: 'WeuiListTips'
+}
+</script>