+++ packages/navbar-item/index.js
@@ -1,0 +1,7 @@
+import WeuiNavbarItem from './src/navbar-item'
+
+WeuiNavbarItem.install = vue => {
+  vue.component(WeuiNavbarItem.name, WeuiNavbarItem)
+}
+
+export default WeuiNavbarItem