+++ packages/navbar/index.js
@@ -1,14 +1,7 @@
 import WeuiNavbar from './src/navbar'
-import WeuiNavbarItem from './src/navbar-item'
 
 WeuiNavbar.install = vue => {
   vue.component(WeuiNavbar.name, WeuiNavbar)
-  vue.component(WeuiNavbarItem.name, WeuiNavbarItem)
-}
-
-export {
-  WeuiNavbar,
-  WeuiNavbarItem
 }
 
 export default WeuiNavbar