+++ packages/preview/src/list.vue
@@ -1,0 +1,10 @@
+<template>
+  <div class="weui-form-preview__list">
+    <slot></slot>
+  </div>
+</template>
+<script>
+export default {
+  name: 'WeuiPreviewList',
+};
+</script>