+++ packages/tab-panel/index.js
@@ -1,0 +1,7 @@
+import WeuiTabPanel from './src/tab-panel'
+
+WeuiTabPanel.install = vue => {
+  vue.component(WeuiTabPanel.name, WeuiTabPanel)
+}
+
+export default WeuiTabPanel