+++ packages/tabbar-item/index.js
@@ -1,0 +1,7 @@
+import WeuiTabbarItem from './src/tabbar-item'
+
+WeuiTabbarItem.install = vue => {
+  vue.component(WeuiTabbarItem.name, WeuiTabbarItem)
+}
+
+export default WeuiTabbarItem