+++ packages/tabbar/index.js
@@ -1,14 +1,7 @@
 import WeuiTabbar from './src/tabbar'
-import WeuiTabbarItem from './src/tabbar-item'
 
 WeuiTabbar.install = vue => {
   vue.component(WeuiTabbar.name, WeuiTabbar)
-  vue.component(WeuiTabbarItem.name, WeuiTabbarItem)
-}
-
-export {
-  WeuiTabbar,
-  WeuiTabbarItem
 }
 
 export default WeuiTabbar